refactor: split worktree.py into core + compat modules
- Extract ManagerCompatMixin to worktree_manager_compat.py (385 lines)
- Reduce worktree.py from 680 to 351 lines (below 450 limit)
- Use TYPE_CHECKING for mixin type safety
- Use NotImplementedError for abstract methods
- Fix test mock paths for append_orchestra_event

Both files now under 450 line limit. Compat methods will be
gradually phased out as architecture evolves.
